+++ packages/liveblocks-react/src/index.ts
@@ -78,6 +78,7 @@ export {
   useHistoryVersionStorageData,
   useHistoryVersionYjsData,
   useDeleteHistoryVersion,
+  useRestoreToStorageVersion,
 } from "./room";
 
 // Export the classic (non-Suspense) versions of our hooks